Add power of spaces
