+++ tests/test_new_space.py
@@ -5,56 +5,11 @@ This should run as part of the CI/CD pipeline.
 
 import pytest
 
-from cadcad.errors import InstanceError
-from cadcad.new_space import space
+from cadcad.builtins import EmptySpace, Integer, Real
+
+# from cadcad.errors import InstanceError
+from cadcad.spaces import space
 
 
 def test_space_creation() -> None:
     """Test creation of spaces."""
-    print()
-
-    @space
-    class Real:
-        value: float
-
-    @space
-    class Integer:
-        value: int
-
-    MySpace = Real.cartesian(Integer)
-    MySpaceMul = Real * Integer
-
-    print(Real)
-    print(type(Real))
-
-    print()
-
-    print(MySpace)
-    print(type(MySpace))
-
-    print()
-
-    print(MySpaceMul)
-    print(type(MySpaceMul))
-
-
-# @space
-# class Real:
-#     value: float
-
-#     def sum():
-#         pass
-
-
-# @space
-# class Real2:
-#     value: Real
-#     value2: Real
-
-# @space
-# class Real2:
-#     value: List[Real]
-
-# Real * Real -> Real^2  # raise warning and rename automatically
-
-# Real.add_dims()
